refactoring img to icons, change directory to UI, change directory ShoppingCart to Product
