+++ src/components/UI/Footer/ArrowUpComponent.tsx
@@ -1,8 +1,8 @@
 // #region Imports
 import Image from 'next/image';
 import React from 'react';
-import ArrowUpBlack from '@/img/ArrowUp(Black).svg';
-import ArrowUpWhite from '@/img/ArrowUp(White).svg';
+import ArrowUpBlack from '@/components/UI/icons/ArrowUp(Black).svg';
+import ArrowUpWhite from '@/components/UI/icons/ArrowUp(White).svg';
 // #endregion
 
 const ArrowUpComponent: React.FC = () => {
